pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-loom"
version = "1.0.0"
description = "A parsimonious research pipeline framework with caching, artifact management, and export capabilities"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Research Team", email = "research@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
data = [
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
    "zstandard>=0.19.0",
]
viz = [
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=0.991",
]
all = [
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
    "zstandard>=0.19.0",
    "matplotlib>=3.5.0",
    "plotly>=5.0.0",
]

[project.scripts]
research-loom = "research_loom.cli:main"

[tool.setuptools]
package-dir = {"research_loom" = "."}
packages = ["research_loom", "research_loom.pipeline"]

